+++ libknet/libknet.h
@@ -760,7 +760,7 @@ int knet_host_set_policy(knet_handle_t knet_h, uint16_t host_id,
  */
 
 int knet_host_get_policy(knet_handle_t knet_h, uint16_t host_id,
-			 int *policy);
+			 uint8_t *policy);
 
 /*
  * knet_host_enable_status_change_notify
